python: Fix setup.py outside of CMake builds
CMake builds declare CMAKE_CURRENT_SOURCE_DIR to be the root of the
repository. py_dir is then determined by joining that path with
"watchman/python". However, when CMAKE_CURRENT_SOURCE_DIR isn't set, we
fallback on setting the source dir to the parent directory of "python",
which then causes a failure because we try to do
"watchman/watchman/python". Presumably the directory structure was
different at some point and nobody updated this.